GP-17 Вторая часть модуля решателя задачи

+++ src/main/java/org/jenjetsu/graduate_project/service/FormulaCalculatorService.java
@@ -1,10 +1,11 @@
 package org.jenjetsu.graduate_project.service;
 
+import java.math.*;
 import java.util.Map;
 
 import org.jenjetsu.graduate_project.entity.formula.*;
 
 public interface FormulaCalculatorService {
     
-    public double calculate(FormulaTreeNode head, Map<String, Object> variableValueMap);
+    public BigDecimal calculate(FormulaTreeNode head, Map<String, Object> variableValueMap);
 }
